+++ tests/blueprints/test_main_config_bp.py
@@ -2,9 +2,8 @@ import os
 from typing import Generator
 
 import pytest
-from flask import Flask, Response, jsonify, make_response
+from flask import Flask, Response
 from flask.testing import FlaskClient
-from pytest import MonkeyPatch
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -49,29 +48,3 @@ def client(
     A Flask test client for the app.
     """
     return app.test_client()
-
-
-def test_good_healthcheck(client: "FlaskClient[Response]") -> None:
-    response = client.get("/health")
-
-    assert response.status_code == 200
-    assert response.content_type == "application/json"
-    expected_json = {"status": "ok"}
-    assert response.get_json() == expected_json
-
-
-def test_bad_healthcheck(
-    client: "FlaskClient[Response]", monkeypatch: MonkeyPatch
-) -> None:
-    def mock_bad_health() -> Response:
-        return make_response(
-            jsonify({"status": "error", "reason": "simulated failure"}), 503
-        )
-
-    monkeypatch.setattr("script_runner.app_blueprint.health", mock_bad_health)
-    response = client.get("/health")
-
-    assert response.status_code == 503
-    assert response.content_type == "application/json"
-    expected_json = {"status": "error", "reason": "simulated failure"}
-    assert response.get_json() == expected_json
